Reverted changes made while adding vite and adjust island size and placement
