+++ package.json
@@ -9,15 +9,19 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@react-three/drei": "^9.57.2",
-    "@react-three/fiber": "^8.12.0",
+    "@react-three/drei": "^9.57.0",
+    "@react-three/fiber": "^8.11.5",
+    "@react-three/postprocessing": "^2.7.0",
+    "@types/three": "^0.149.0",
+    "leva": "^0.9.34",    
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "three": "^0.150.1"
+    "react-spring": "^9.7.1",
+    "three": "^0.150.1",
+    "use-gesture": "^1.0.0",
+    "web-vitals": "^2.1.4"
   },
   "devDependencies": {
-    "@types/react": "^18.0.27",
-    "@types/react-dom": "^18.0.10",
     "@vitejs/plugin-react": "^3.1.0",
     "prettier": "2.8.4",
     "r3f-perf": "^7.1.2",
