+++ src/index.jsx
@@ -1,5 +1,24 @@
 import React from "react";
-import ReactDOM from "react-dom/client";
-import App from "./App";
+import "./style.css";
+import { createRoot } from "react-dom/client";
+import { Canvas } from "@react-three/fiber";
+import { OrbitControls } from "@react-three/drei";
+import { Perf } from "r3f-perf";
+import { Experience } from "./Experience";
 
-ReactDOM.createRoot(document.getElementById("root")).render(<App />);
+function App() {
+  return (
+    <Canvas>
+      <color args={["#1E1E1E"]} attach="background" />
+      <Perf />
+      <OrbitControls makeDefault />
+      <ambientLight intensity={0.1} />
+      <directionalLight />
+      <Experience />
+    </Canvas>
+  );
+}
+
+const root = createRoot(document.getElementById("root"));
+
+root.render(<App />);
